Machine Learning/scripts/requirements.txt
paho-mqtt
firebase-admin
joblib
pandas
numpy
scikit-learn
lightgbm
